refactor: removed unnecessary variables, made base64 module
