+++ apps/shopify/src/collectionPagination.js
@@ -1,6 +1,6 @@
 import { collectionDataTransformer } from './dataTransformer';
 import BasePagination from './basePagination';
-import checkAndConvertToBase64 from './utils/checkAndConvertToBase64'
+import { checkAndConvertToBase64 } from './utils/base64'
 
 const makePagination = async (sdk) => {
   const pagination = new BasePagination({
@@ -9,16 +9,14 @@ const makePagination = async (sdk) => {
     fetchProducts: async function (search, PER_PAGE) {
       const query = { query: search };
 
-      const response = await this.shopifyClient.collection.fetchQuery({
+      const collections = await this.shopifyClient.collection.fetchQuery({
         first: PER_PAGE,
         sortBy: 'TITLE',
         reverse: true,
         ...(search.length && query),
       });
 
-      const collections = response.map((res) => { return checkAndConvertToBase64(res) })
-
-      return collections
+      return collections.map((collection) => checkAndConvertToBase64(collection))
 
     },
   });
