--- apps/shopify/src/utils/atob.js
@@ -1,8 +1,0 @@
-export default function atob(str) {
-    try {
-        const unencodedId = window.atob(str)
-        return unencodedId
-    } catch (error) {
-        return null
-    }
-}
