--- apps/shopify/src/utils/btoa.js
@@ -1,8 +1,0 @@
-export default function atob(str) {
-    try {
-        const decodedId = window.btoa(str)
-        return decodedId
-    } catch (error) {
-        return null
-    }
-}
