--- apps/shopify/src/utils/isBase64.js
@@ -1,9 +1,0 @@
-export default function isBase64(str) {
-
-    if (str === '' || str.trim() === '') { return false; }
-    try {
-        return window.btoa(window.atob(str)) === str;
-    } catch (err) {
-        return false;
-    }
-}
